+++ backend/JiShop/src/main/java/com/jishop/member/dto/request/FindUserRequest.java
@@ -1,0 +1,6 @@
+package com.jishop.member.dto.request;
+
+public record FindUserRequest(
+        String phone
+) {
+}
