+++ backend/JiShop/src/main/java/com/jishop/member/service/UserService.java
@@ -5,6 +5,8 @@ import com.jishop.member.dto.SignInFormRequest;
 import com.jishop.member.dto.SignUpFormRequest;
 import com.jishop.member.dto.SocialUserInfo;
 import com.jishop.member.dto.Step1Request;
+import com.jishop.member.dto.request.FindUserRequest;
+import com.jishop.member.dto.response.FindUserResponse;
 import jakarta.servlet.http.HttpSession;
 
 public interface UserService {
@@ -14,4 +16,5 @@ public interface UserService {
     void signIn(SignInFormRequest form, HttpSession session);
     void emailcheck(Step1Request request);
     String loginStr(Long userId);
+    FindUserResponse findUser(FindUserRequest request);
 }
